+++ R/haplotype_refinement.R
@@ -1,6 +1,6 @@
 # Author: Rensc
 # Date: 2026-07-30
-# Version: dev005
+# Version: dev006
 # Function: Refine haplotypes and prioritize phenotype-associated variant effects
 # Input: HapVariant objects and phenotype tables
 # Output: Refined haplotype objects, variant-effect tables, and ggplot figures
@@ -38,9 +38,15 @@
 #' pheno <- read_pheno(pheno_file, verbose = FALSE)
 #' anno <- read_genepred(anno_file, format = "genePredExt", verbose = FALSE)
 #' hap <- hap_variant(vcf, annotation = anno, gene_id = "GeneA", genotype_mode = "string", min_variant_number = 1)
-#' refined <- refine_haplotype(hap, phenotype = pheno, traits = "seed_weight", min_hap_samples = 3)
+#' refined <- refine_haplotype(
+#'   hap,
+#'   phenotype = pheno,
+#'   traits = "protein_content",
+#'   min_hap_samples = 3,
+#'   effect_threshold = 0.5
+#' )
 #' refined$refined_haplotypes
-#' plot_hap_pheno(refined$refined_hap, phenotype = pheno, traits = "seed_weight", min_hap_samples = 3)
+#' plot_hap_pheno(refined$refined_hap, phenotype = pheno, traits = "protein_content", min_hap_samples = 3)
 #' @export
 refine_haplotype <- function(hap,
                              phenotype,
@@ -269,8 +275,14 @@ refine_haplotype <- function(hap,
 #' pheno <- read_pheno(pheno_file, verbose = FALSE)
 #' anno <- read_genepred(anno_file, format = "genePredExt", verbose = FALSE)
 #' hap <- hap_variant(vcf, annotation = anno, gene_id = "GeneA", genotype_mode = "string", min_variant_number = 1)
-#' refined <- refine_haplotype(hap, phenotype = pheno, traits = "seed_weight", min_hap_samples = 3)
-#' plot_refined_hap_pheno(refined, phenotype = pheno, traits = "seed_weight", min_hap_samples = 3)
+#' refined <- refine_haplotype(
+#'   hap,
+#'   phenotype = pheno,
+#'   traits = "protein_content",
+#'   min_hap_samples = 3,
+#'   effect_threshold = 0.5
+#' )
+#' plot_refined_hap_pheno(refined, phenotype = pheno, traits = "protein_content", min_hap_samples = 3)
 #' @export
 plot_refined_hap_pheno <- function(refined_hap,
                                    phenotype,
@@ -401,7 +413,13 @@ plot_refined_hap_pheno <- function(refined_hap,
 #' pheno <- read_pheno(pheno_file, verbose = FALSE)
 #' anno <- read_genepred(anno_file, format = "genePredExt", verbose = FALSE)
 #' hap <- hap_variant(vcf, annotation = anno, gene_id = "GeneA", genotype_mode = "string", min_variant_number = 1)
-#' refined <- refine_haplotype(hap, phenotype = pheno, traits = "seed_weight", min_hap_samples = 3)
+#' refined <- refine_haplotype(
+#'   hap,
+#'   phenotype = pheno,
+#'   traits = "protein_content",
+#'   min_hap_samples = 3,
+#'   effect_threshold = 0.5
+#' )
 #' plot_refined_hap_variant(refined, annotation = anno, min_hap_samples = 1)
 #' @export
 plot_refined_hap_variant <- function(refined_hap,